Stop using an emulated get operation
Stop using an emulated get operation (which actually uses the
open/read/close API to stream data and a stream checksum check, behind
the scenes) and use the real get API.

This allows us to have a new --verify option on baton-get/baton-do to
enable full, at-rest checksums when downloading to files (i.e. when
using --save), using get's built-in check to do this (currently
on-the-fly checksums, not at-rest).

As get uses parallel chunk download for large files (>32MiB), that
part of the operation is likely to be faster than before. However,
at-rest checksums will take additional time, so there will be a
penalty there.

If the performance penalty of at-rest checksums is not acceptable,
avoiding use of the --verify flag should retain current speed. In
doing so, we will lose the on-the-fly checksum. However, an at-rest
checksum is by far the more preferable option for data integrity and
the effectiveness of the on-the-fly checksum is debatable.

There will be a new --no-clobber option to enable get's local file
protection, but baton's default of clobbering local files will remain,
so that pipelines using it are idempotent by default. Pipelines
requiring local file protection can enable --no-clobber and then
become responsible for cleaning/re-trying failed downloads their own
way.